# frontend/src/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/projects":{"get":{"summary":"Get Projects","operationId":"get_projects_api_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Project"},"type":"array","title":"Response Get Projects Api Projects Get"}}}}}}},"/api/project/{project_slug}":{"get":{"summary":"Get Project","operationId":"get_project_api_project__project_slug__get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/calculate":{"get":{"summary":"Get Project","operationId":"get_project_api_project__project_slug__calculate_get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/videos":{"get":{"summary":"Get Videos","operationId":"get_videos_api_project__project_slug__videos_get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Video"},"title":"Response Get Videos Api Project  Project Slug  Videos Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/video/{video_slug}":{"get":{"summary":"Get Video","operationId":"get_video_api_project__project_slug__video__video_slug__get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}},{"name":"video_slug","in":"path","required":true,"schema":{"type":"string","title":"Video Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Video"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/final":{"get":{"summary":"Build Final Cut","operationId":"build_final_cut_api_project__project_slug__final_get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/video/{video_slug}/segments":{"post":{"summary":"Set Video Segments","operationId":"set_video_segments_api_project__project_slug__video__video_slug__segments_post","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}},{"name":"video_slug","in":"path","required":true,"schema":{"type":"string","title":"Video Slug"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Segment"},"title":"Segments"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Video"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/video/{video_slug}/thumbnail":{"get":{"summary":"Get Video Segments","operationId":"get_video_segments_api_project__project_slug__video__video_slug__thumbnail_get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}},{"name":"video_slug","in":"path","required":true,"schema":{"type":"string","title":"Video Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/project/{project_slug}/video/{video_slug}/preview":{"get":{"summary":"Get Video Preview","operationId":"get_video_preview_api_project__project_slug__video__video_slug__preview_get","parameters":[{"name":"project_slug","in":"path","required":true,"schema":{"type":"string","title":"Project Slug"}},{"name":"video_slug","in":"path","required":true,"schema":{"type":"string","title":"Video Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{filepath}":{"get":{"summary":"Serve React App","operationId":"serve_react_app__filepath__get","parameters":[{"name":"filepath","in":"path","required":true,"schema":{"type":"string","title":"Filepath"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InterestLevel":{"properties":{"timestamp":{"type":"number","title":"Timestamp"},"interest_level":{"type":"number","title":"Interest Level"}},"type":"object","required":["timestamp","interest_level"],"title":"InterestLevel"},"Project":{"properties":{"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"},"videos":{"items":{"$ref":"#/components/schemas/Video"},"type":"array","title":"Videos"}},"type":"object","required":["name","slug","videos"],"title":"Project"},"Segment":{"properties":{"start_time":{"type":"number","title":"Start Time"},"end_time":{"type":"number","title":"End Time"}},"type":"object","required":["start_time","end_time"],"title":"Segment"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Video":{"properties":{"project_dir_name":{"type":"string","title":"Project Dir Name"},"length":{"type":"number","title":"Length"},"size_bytes":{"type":"integer","title":"Size Bytes"},"slug":{"type":"string","title":"Slug"},"mp4_filename":{"type":"string","title":"Mp4 Filename"},"lrv_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lrv Filename"},"thumbnail_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Filename"},"accel_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Accel Filename"},"gyro_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gyro Filename"},"segments_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Segments Filename"},"suggested_segments":{"anyOf":[{"items":{"$ref":"#/components/schemas/Segment"},"type":"array"},{"type":"null"}],"title":"Suggested Segments"},"interest_levels":{"anyOf":[{"items":{"$ref":"#/components/schemas/InterestLevel"},"type":"array"},{"type":"null"}],"title":"Interest Levels"},"accel":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Accel","default":[]},"gyro":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Gyro","default":[]},"segments":{"items":{"$ref":"#/components/schemas/Segment"},"type":"array","title":"Segments","default":[]}},"type":"object","required":["project_dir_name","length","size_bytes","slug","mp4_filename","lrv_filename","thumbnail_filename","accel_filename","gyro_filename","segments_filename"],"title":"Video"}}}}